Initial commit
This is a template for how I think LoveOS gems should look. It's got it all - a
service, a client, and a command-line interface which can start a server or use
the client. It's also packageable as a gem.

On the other hand, it's chunky bacon. It doesn't do much. Still there are enough
best practices here to make it worth looking into from time to time.